pyproject.toml
[tool.poetry]
name = "inkplate-dashboard"
version = "0.1.0"
description = "Project to display news and weather on an Inkplate 10"
authors = ["Ugo Méda <meda.ugo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tomlkit = "^0.13.2"
Jinja2 = "^3.1.4"
requests = "^2.32.3"
pydantic = "^2.9.2"
pillow = "^11.0.0"
defusedxml = "^0.7.1"
babel = "^2.16.0"
libsass = "^0.23.0"
starlette = "^0.41.0"
uvicorn = "^0.32.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"
mypy = "^1.12.0"
types-requests = "^2.32.0.20241016"
types-defusedxml = "^0.7.0.20240218"
types-libsass = "^0.23.0.20240724"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = [
    "F",
    "E", "W",
    "YTT",
    "B",
    "Q",
    "G",
    "C4",
    "PLE", "PLW",
    "PIE",
    "TID",
    "ISC",
    "TCH",
    "I",
    "UP",
]
ignore = ["ISC001"]

[tool.mypy]
python_version = "3.11"
strict = true
